requirements.txt
# ==================== CORE DEPENDENCIES ====================
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
groq>=0.4.0
python-dotenv>=1.0.0

# ==================== RAG & LLM INTEGRATION ====================
langchain>=0.1.0
langchain-community>=0.0.1
langchain-pinecone>=0.0.1
sentence-transformers>=2.2.2
huggingface-hub>=0.17.0

# ==================== TRANSCRIPTION & NLP ====================
openai-whisper>=20240101
spacy>=3.7.0
langdetect>=1.0.9
transformers>=4.30.0

# ==================== DATA PROCESSING ====================
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
regex>=2023.0

# ==================== REAL-TIME & ASYNC ====================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
aiohttp>=3.9.0

# ==================== DATABASE & STORAGE ====================
pinecone-client>=2.2.0
redis>=5.0.0

# ==================== MONITORING & LOGGING ====================
python-json-logger>=2.0.0

# ==================== NETWORK & HTTP ====================
requests>=2.31.0
httpx>=0.25.0

# ==================== DEVELOPMENT & TESTING ====================
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=23.10.0
flake8>=6.1.0
mypy>=1.7.0
watchdog>=3.0.0